chore: Update secrets configuration for vaultwarden service
This commit updates the secrets configuration for the vaultwarden service in the greyserver module. It removes the individual secret entries for "vaultwarden/admin_token" and "vaultwarden/db_pass" and replaces them with a single entry for "vaultwarden". The corresponding changes are made in the sops.nix and vaultwarden.nix files.

Signed-off-by: Scott Gould <greysilly7@gmail.com>