chore: Update hardware-configuration.nix
This commit updates the hardware-configuration.nix file by adding options for the "/boot" file system. The options "fmask=0022" and "dmask=0022" are included to set the file and directory permissions. Additionally, the swapDevices list is cleared, removing any swap devices. These changes are made to improve the system's boot process and optimize resource allocation.

Signed-off-by: Scott Gould <greysilly7@gmail.com>