chore: Update configuration files for secureboot and sops
This commit updates the configuration files for secureboot and sops in the greypersonal and greyserver hosts. The secureboot.nix and sops.nix files are deleted from their respective hosts, and the imports in the configuration.nix files are modified to use the new module paths. This change is made to improve the organization and maintainability of the configuration files.

Signed-off-by: Scott Gould <greysilly7@gmail.com>