feat: Update spacebar service configuration and storage location
The spacebar service configuration in the flake.nix and spacebar.nix files is updated to include the correct path for starting the spacebar bundle. Additionally, a new storage location at /var/lib/spacebar is created and assigned appropriate permissions. This change ensures the proper execution and storage of the Spacebar service.

Signed-off-by: Scott Gould <greysilly7@gmail.com>