chore: Update spacebar service configuration and storage location
This commit updates the spacebar service configuration in the flake.nix and spacebar.nix files to include the correct path for starting the spacebar bundle. It also creates a new storage location at /var/lib/spacebar with appropriate permissions. These changes ensure the proper execution and storage of the Spacebar service.

Signed-off-by: Scott Gould <greysilly7@gmail.com>