feat: Add nh.nix module for configuring nh program
This commit adds a new module, nh.nix, which allows configuring the nh program. The nh program is enabled with the clean feature and additional arguments are provided to specify the cleanup behavior. The flake path is set to "/home/user/my-nixos-config". This module is included in the greypersonal configuration.

Signed-off-by: Scott Gould <greysilly7@gmail.com>