feat: Add tabby service configuration
The code changes include adding the `tabby.nix` module to the `greyserver` configuration. This module is responsible for configuring the tabby service.

Recent user commits and repository commits indicate a pattern of updating service configurations, including the spacebar service. Therefore, it is consistent to add a commit message for updating the tabby service configuration.

Signed-off-by: Scott Gould <greysilly7@gmail.com>