Update adguard.nix to include an additional user in the users list
The code changes modify the adguard.nix file to add a new user to the users list. The new user has the name "admin" and a hashed password. This change is made to enhance the security of the AdGuard service.

Recent user commits and repository commits indicate a pattern of updating service configurations, including the tabby and spacebar services. Therefore, it is consistent to add a commit message for updating the adguard service configuration.

Signed-off-by: Scott Gould <greysilly7@gmail.com>