feat: Add hyprlock and hypridle to startup configuration
The code changes include adding `hyprlock` and `hypridle` to the startup configuration in the `config.nix` file. This will ensure that `hyprlock` and `hypridle` are launched when the system starts up.

Based on the recent user commits and repository commits, it seems that the developer has been working on various improvements and fixes. However, none of the recent commits directly relate to the changes made in this code.

Please note that the commit message should be adjusted according to the repository's commit message conventions and guidelines.

Signed-off-by: Scott Gould <greysilly7@gmail.com>