chore: Update secrets encryption recipients
The code changes in `.sops.yaml` and `secrets/secrets.yaml` update the recipients for encryption in the secrets files. The recipient for the `computer_laptop` key has been changed to `age1kwvkrkzrtwg66yzjn6hlhxj6dyguxn2vypumpr06a6ypf49cf3gswgrrnk`. This change ensures that the secrets are encrypted and accessible only to the intended recipients.

Signed-off-by: Scott Gould <greysilly7@gmail.com>