chore: Update impermanence.nix to rollback BTRFS root subvolume
This commit updates the impermanence.nix file to add a systemd service for rolling back the BTRFS root subvolume to a pristine state. The service is configured to run after the systemd-cryptsetup@crypted.service and before the sysroot.mount. It mounts the BTRFS filesystem, deletes all subvolumes except /root, restores a blank /root subvolume, and then unmounts the filesystem. This improves the rollback functionality and ensures a clean state for the root subvolume.

Signed-off-by: Scott Gould <greysilly7@gmail.com>