feat: Add support for serving images from a specific directory
The code changes in `spacebarserver.nix` introduce a new configuration for serving images from a specific directory. This change adds a new route `/images` with the root directory set to `${inputs.spacebarchat.packages.${"x86_64-linux"}.default}/images`. This will allow the server to serve images from the specified directory.

Note: The recent user and repository commits are not directly related to the code changes and do not provide any additional context for the commit message.
Signed-off-by: Scott Gould <greysilly7@gmail.com>