chore: Update build.yaml to use flake for nix-build
The code changes in `.github/workflows/build.yaml` modify the build workflow to use the `nix-build --flake .#greypersonal` command instead of `nix-build`. This change allows for building the project using the specified flake configuration.

Note: The recent user and repository commits are not directly related to the code changes and do not provide any additional context for the commit message.
Signed-off-by: Scott Gould <greysilly7@gmail.com>