Update default.nix and jankclient.nix
This commit updates the default.nix file in the hosts directory and the jankclient.nix file in the system/server directory.

In default.nix, the changes include adding new configurations for the WSL (Windows Subsystem for Linux) setup, enabling WSL and using Windows drivers. The system state version is also updated to 24.05 and the host name is set to "greyworkstation".

In jankclient.nix, the commit updates the sha256 value for the JankClient repository.

These changes are made to improve the configuration and ensure compatibility with the latest versions.

Signed-off-by: Scott Gould <greysilly7@gmail.com>